GH-32 fix the 2 Ingnored test and change all the Float into Double

+++ src/main/java/fr/unice/polytech/si4/otake/cookiefactory/RecipeBook.java
@@ -9,19 +9,19 @@ import fr.unice.polytech.si4.otake.cookiefactory.product.cookie.Cookie;
 
 public class RecipeBook {
 
-	private Map<Cookie, Float> allCookies;
+	private Map<Cookie, Double> allCookies;
 
 	/**
 	 * new RecipeBook to manage the Cookies recipes
 	 */
 	public RecipeBook() {
-		this(new HashMap<Cookie, Float>());
+		this(new HashMap<Cookie, Double>());
 	}
 
 	/**
 	 * new RecipeBook to manage the Cookies recipes
 	 */
-	public RecipeBook(Map<Cookie, Float> maps) {
+	public RecipeBook(Map<Cookie, Double> maps) {
 		this.allCookies = maps;
 	}
 
@@ -32,7 +32,7 @@ public class RecipeBook {
 	 * @return the cookie if it exist, null otherwise
 	 */
 	public Cookie getCookie(String name) {
-		for (Map.Entry<Cookie, Float> entry : allCookies.entrySet()) {
+		for (Map.Entry<Cookie, Double> entry : allCookies.entrySet()) {
 			Cookie cookie = entry.getKey();
 			if (name != null && name.equals(cookie.getName())) {
 				return cookie;
@@ -48,7 +48,7 @@ public class RecipeBook {
 	 */
 	public List<Cookie> getCookies() {
 		List<Cookie> cookiesreturn = new ArrayList<>(allCookies.size());
-		for (Map.Entry<Cookie, Float> entry : allCookies.entrySet()) {
+		for (Map.Entry<Cookie, Double> entry : allCookies.entrySet()) {
 			if (!entry.getKey().isCustom()) {
 				cookiesreturn.add(entry.getKey());
 			}
@@ -63,7 +63,7 @@ public class RecipeBook {
 	 * @return false if the cookie already exist, true otherwise
 	 */
 	public boolean addRecipe(Cookie recipe) {
-		return this.allCookies.putIfAbsent(recipe, (float) 0.0) == null;
+		return this.allCookies.putIfAbsent(recipe, 0.0) == null;
 	}
 
 	/**
@@ -76,17 +76,17 @@ public class RecipeBook {
 		return this.allCookies.remove(recipe) != null;
 	}
 
-	public Map<Cookie, Float> getStatistic() {
+	public Map<Cookie, Double> getStatistic() {
 		int soldSum = 0;
-		for (Map.Entry<Cookie, Float> entry : allCookies.entrySet()) {
+		for (Map.Entry<Cookie, Double> entry : allCookies.entrySet()) {
 			soldSum += entry.getKey().getUnits();
 		}
-		for (Map.Entry<Cookie, Float> entry : allCookies.entrySet()) {
+		for (Map.Entry<Cookie, Double> entry : allCookies.entrySet()) {
 			if (soldSum > 0) {
-				Float statistic = ((entry.getKey().getUnits() / (float) soldSum) * 100);
+				Double statistic = ((entry.getKey().getUnits() / soldSum) * 100.);
 				allCookies.put(entry.getKey(), statistic);
 			} else {
-				Float statistic = (float) 0;
+				Double statistic = 0.;
 				allCookies.put(entry.getKey(), statistic);
 			}
 		}
